requirements.txt
# LangGraph与LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# MinerU (PDF解析) - 正确的安装命令
mineru[core]>=0.7.0

# Web框架
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart

# 数据模型与配置
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 工具库
python-dotenv>=1.0.0
loguru>=0.7.0

# 科学计算 (用于Auditor的相似度计算)
numpy>=1.24.0
scikit-learn>=1.3.0

# 数据库 (SQLite持久化)
sqlalchemy>=2.0.0
